fed_macro_mvp/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pypdf>=4.2.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
ollama>=0.3.0
pyarrow>=15.0.0
scikit-learn>=1.4.0
json-repair>=0.30.0
